+++ tests/jsonpickle_test.py
@@ -346,6 +346,13 @@ def test_reduce_with_invalid_data(value, unpickler):
     assert result == []
 
 
+@pytest.mark.parametrize('value', ['', 'x', 1, True, [], {}])
+def test_restore_id_with_invalid_data(value, unpickler):
+    """Invalid serialized ID data results in None"""
+    result = unpickler.restore({'ref': {tags.ID: value}})
+    assert result['ref'] is None
+
+
 def test_dict(pickler, unpickler):
     """Our custom keys are preserved when user dicts contain them"""
     dict_a = {'key1': 1.0, 'key2': 20, 'key3': 'thirty', tags.JSON_KEY + '6': 6}
@@ -604,6 +611,21 @@ def test_restore_legacy_builtins():
     assert ae is AssertionError
     cls = jsonpickle.decode('{"py/type": "__builtin__.int"}')
     assert cls is int
+
+
+@pytest.mark.parametrize(
+    'value,expect',
+    [
+        ('module_does_not_exist/ignored', None),
+        ('builtins/int', None),
+        ('builtins/invalid.int', None),
+        ('builtins/builtinsx.int', None),
+    ],
+)
+def test_restore_invalid_repr(value, expect, unpickler):
+    """Test restoring invalid repr tags"""
+    result = unpickler.restore({tags.REPR: value})
+    assert result is expect
 
 
 def test_unpickler_on_missing():