+++ CHANGES.rst
@@ -1,6 +1,7 @@
 Upcoming
 ========
-    * The unpickler is now more resilient to malformed "py/id" data. (+546)
+    * The unpickler is now more resilient to malformed "py/id" and "py/repr" data.
+      (+546)
 
 v4.0.1
 ======