+++ CHANGES.rst
@@ -1,3 +1,8 @@
+Upcoming
+========
+    * The unpickler is now more resilient to malformed "py/id" and "py/repr" data.
+      (+546)
+
 v4.0.1
 ======
     * The unpickler is now more resilient to malformed "py/reduce", "py/set",