+++ CHANGES.rst
@@ -1,5 +1,5 @@
-Upcoming
-========
+v4.0.2
+======
     * The unpickler is now more resilient to malformed "py/id" and "py/repr" data.
       (+546)
     * The unpickler is now more resilient to invalid "py/b85" and "py/b64" data.